Extract `CapturingKoogHttpClient` to `test-utils` and update tests
