+++ prompt/prompt-executor/prompt-executor-clients/prompt-executor-google-client/src/jvmTest/kotlin/ai/koog/prompt/executor/clients/google/GooglePrimaryConstructorTest.kt
@@ -1,17 +1,14 @@
 package ai.koog.prompt.executor.clients.google
 
-import ai.koog.http.client.KoogHttpClient
 import ai.koog.prompt.dsl.prompt
 import ai.koog.prompt.executor.clients.google.models.GooglePart
 import ai.koog.prompt.executor.clients.google.models.GoogleRequest
 import ai.koog.prompt.executor.clients.google.models.GoogleResponse
 import ai.koog.prompt.llm.LLMProvider
 import ai.koog.prompt.message.Message
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.emptyFlow
+import ai.koog.test.utils.CapturingKoogHttpClient
 import kotlinx.coroutines.test.runTest
 import kotlinx.serialization.json.Json
-import kotlin.reflect.KClass
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertIs
@@ -51,7 +48,7 @@ class GooglePrimaryConstructorTest {
 
     @Test
     fun `primary constructor should execute through koog http client`() = runTest {
-        val transport = CapturingKoogHttpClient { _ ->
+        val transport = CapturingKoogHttpClient(clientName = "CapturingGoogleClient") { _ ->
             json.decodeFromString<GoogleResponse>(responseJson)
         }
         val client = GoogleLLMClient(
@@ -72,42 +69,5 @@ class GooglePrimaryConstructorTest {
         assertEquals("Hello?", userPart.text)
         val message = assertIs<Message.Assistant>(responses.single())
         assertEquals("Hello from Google KoogHttpClient", message.content)
-    }
-
-    private class CapturingKoogHttpClient(
-        private val responder: (KClass<*>) -> Any
-    ) : KoogHttpClient {
-        override val clientName: String = "CapturingGoogleClient"
-        var lastPath: String? = null
-        var lastRequest: Any? = null
-
-        override suspend fun <R : Any> get(path: String, responseType: KClass<R>, parameters: Map<String, String>): R {
-            error("GET is not expected in this test")
-        }
-
-        override suspend fun <T : Any, R : Any> post(
-            path: String,
-            request: T,
-            requestBodyType: KClass<T>,
-            responseType: KClass<R>,
-            parameters: Map<String, String>
-        ): R {
-            lastPath = path
-            lastRequest = request
-            @Suppress("UNCHECKED_CAST")
-            return responder(responseType) as R
-        }
-
-        override fun <T : Any, R : Any, O : Any> sse(
-            path: String,
-            request: T,
-            requestBodyType: KClass<T>,
-            dataFilter: (String?) -> Boolean,
-            decodeStreamingResponse: (String) -> R,
-            processStreamingChunk: (R) -> O?,
-            parameters: Map<String, String>
-        ): Flow<O> = emptyFlow()
-
-        override fun close() = Unit
     }
 }
